Imported Upstream version 0.8~beta1

+++ src/client-api/src/c++/common/syncml/formatter/Formatter.cpp
@@ -1081,7 +1081,7 @@ StringBuffer* Formatter::getSync(Sync* sync) {
         numberOfChanges->sprintf(
             "<%s>%d</%s>",
             NUMBER_OF_CHANGES,
-            sync->getNumberOfChanges(),
+            (int)sync->getNumberOfChanges(),
             NUMBER_OF_CHANGES);
     }
 
@@ -1090,7 +1090,8 @@ StringBuffer* Formatter::getSync(Sync* sync) {
     sequence = getSpecificCommand(sync->getCommands(), SEQUENCE);
     atomic   = getSpecificCommand(sync->getCommands(), ATOMIC);
 
-    if (NotZeroStringBufferLenght(10, cmdID, cred, meta, source, target, noResp, numberOfChanges, commonCommandList, sequence, atomic)) {
+    if (NotZeroStringBufferLenght(10, cmdID, cred, meta, source, target, noResp,
+                                  numberOfChanges, commonCommandList, sequence, atomic)) {
         s = new StringBuffer();
         s->append(cmdID);
         s->append(noResp);
@@ -1105,7 +1106,8 @@ StringBuffer* Formatter::getSync(Sync* sync) {
     }
 
     ret = getValue(SYNC, s);
-    deleteAllStringBuffer(11, &s, &cred, &cmdID, &meta, &source, &target, &noResp, &numberOfChanges, &commonCommandList, &atomic, &sequence);
+    deleteAllStringBuffer(11, &s, &cred, &cmdID, &meta, &source, &target, &noResp,
+                          &numberOfChanges, &commonCommandList, &atomic, &sequence);
 
     return ret;
 }
